apps/api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
supabase==2.5.0
PyJWT==2.8.0
google-generativeai==0.7.1
google-cloud-pubsub==2.23.1
google-cloud-bigquery==3.26.0
google-cloud-secret-manager==2.21.0
redis==5.0.7
pytest==8.3.3
pytest-asyncio==0.23.7
httpx==0.27.0
